feat: add tilemap colliders and jump buffering
